by-value model for header fields, diag router test working
I ran a number of tests today on code that carefully manages references
to avoid copying headers through a pipeline and code that just says fuck
it and passes everything by value leaving it up to the rust compiler to
try to optimize data movement as much as possible. It turns out the
latter is not only much simpler code, it's also at least 2x faster in my
little benchmarks - so going with by-value header members for now.

Also start a few background/design docs.