ensure correct table key width, more probes
Many table key types are stored as BigUint types. This presents an
opportunity for key width to be lost. For example, a value of 47 encoded
as 2 bytes can be squashed down to 1 byte with BigUint. This commit adds
metadata to table keys to track key width.

Also add a p4rs dtrace probe for table misses to help provide
debugability to the common question of why are my packets being dropped
due to table misses.