build a workable softnpu test apparatus, shore up tests
The SoftNpu implementation in the test directory was a really leaky
abstraction. It required the test writer to be intimately familiar with
xfr mechanics, was extremely verbose, and error prone to use in terms
of setting up phys, rings, pipelines etc.

This commit replaces the SoftNpu implementation in the test directory.
The new implementation that makes it much easier and less error prone to
write tests. There is also a blocking I/O interface now, which
eliminates the need for inserting sleeps in CI tests to ensure packets
have propagated.

This new SoftNpu interface also provides the basis for providing
runnable documentation for users.