fix double reference for bit/varbit/int types
We need to ensure that bit/varbit/int types are passed by reference, but
there are other conditions that can lead to the variable being
referenced, so we need to make sure not to double reference the
variable.

This commit moves the referencing logic further down the argument
construction logic to only create reference outside the direciton
out/inout path as that path is already adding mutable references to
arguments.