recombine parse/ingress/egress into process_packet
Breaking out the pipeline trait into distinct methods for parsing,
ingress and egress created an issue of how to handle `x4c` compiler
generated types that are only really for internal pipeline
implementation and not for external consumption. The separating of the
parse/ingress/egress methods into distinct entitites exposed these
compiler/pipeline internals to consumers of the Pipeline trait and there
was no real good way to handle this beyond exposing these types as
Box<dyn Foo> objects and comming up with a minimum viable Foo trait for
each pipeline-internal type.

That, however, is counter to the point of these data structures as
compiler/pipeline internal. Because they exit the pipeline through
return values and the caller has to plumb them back in to the other
functions as parameters, the pipeline then needs to cast them back into
concreate types in order to be useful in the pipeline execution context.
So this is a lose-lose situation. Callers have to shlep opaque-ish types
around and pipelines have to down-cast opaque-ish types to something
useful.

This commit is still keeping the ingress/egress breakout distinction.
But the mechanics of executing ingress, egress and the possible packet
replication logic between is now handled internally by the generated
pipeline itself.

In sum this results for a small net-change for consumers of the Pipeline
trait. Just the return value of `process_packet` changes. Instead of
returning an `Option<(packet_out<'a>, u16)>` we return a
`Vec<(packet_out<'a>, u16)>` to enable a single packet to potentially
replicate into a set of packets.